// src/components/box/enum.ts
export enum BoxFillState {
	PRIMARY = "bg-primary-400",
	SECONDARY = "bg-secondary",
	GREEN = "bg-green",
	PINK = "bg-pink",
	TERTIARY = "bg-neutral-100",
	NONE = "bg-transparent",
}

export enum BoxOutlineState {
	PRIMARY = "border border-primary-400",
	SECONDARY = "border border-secondary",
	GREEN = "border border-green",
	PINK = "border border-pink",
	TERTIARY = "border border-neutral-100",
	NONE = "border-0",
}